Update README.md
Build an app or a series of apps that run identically on AWS, Azure, and GCS in Windows and Linux to learn cloud differences. Use benchmark style apps (e.g. SuperPi) to push the limits of cloud capabilities to determine limits of cloud offerings and effectiveness of auto-scaling and redundancy features. Using python because of its support across Linux and Windows and its support across cloud service providers.